Separate cloud/non-cloud tests
