+++ test/functional-cloud/datastoresrp/resources/testdata/datastoresrp-resources-microsoft-sql.bicep
