+++ test/functional-cloud/datastoresrp/resources/testdata/datastoresrp-resources-mongodb-recipe-context.bicep
