+++ test/functional-cloud/datastoresrp/resources/testdata/datastoresrp-resources-redis-default-recipe.bicep
