+++ test/functional-cloud/datastoresrp/resources/testdata/datastoresrp-resources-redis-manual.bicep
