+++ test/functional-cloud/datastoresrp/resources/testdata/datastoresrp-resources-redis-recipe.bicep
