+++ test/functional-cloud/shared/mechanics/testdata/aws-mechanics-redeploy-withupdatedresource.step1.bicep
