+++ test/functional-cloud/shared/mechanics/testdata/aws-mechanics-redeploy-withupdatedresource.step2.bicep
