+++ test/functional-cloud/shared/resources/aws_s3_bucket_test.go
