+++ test/functional-cloud/shared/resources/azure_connections_test.go
