+++ test/functional-cloud/shared/resources/storage_test.go
