+++ test/functional-cloud/shared/resources/testdata/aws-multi-identifier.bicep
