+++ test/functional-cloud/shared/resources/testdata/aws-s3-bucket.bicep
