+++ test/functional-cloud/shared/resources/testdata/corerp-azure-connection-database-service.bicep
@@ -1,0 +1,31 @@
+import radius as radius
+
+param magpieimage string
+
+param environment string
+
+param cosmosmongodbresourceid string
+
+resource app 'Applications.Core/applications@2023-10-01-preview' = {
+  name: 'corerp-azure-connection-database-service'
+  location: 'global'
+  properties: {
+    environment: environment
+  }
+}
+
+resource store 'Applications.Core/containers@2023-10-01-preview' = {
+  name: 'db-service'
+  location: 'global'
+  properties: {
+    application: app.id
+    container: {
+      image: magpieimage
+    }
+    connections: {
+      databaseresource: {
+        source: cosmosmongodbresourceid
+      }
+    }
+  }
+}
