+++ test/functional-cloud/shared/resources/testdata/corerp-resources-extender-aws-s3-recipe.bicep
