+++ test/functional-cloud/shared/resources/testdata/corerp-resources-terraform-azurestorage.bicep
