+++ test/functional-cloud/shared/resources/testdata/corerp-resources-volume-azure-keyvault.bicep
