+++ test/functional-cloud/shared/resources/testdata/recipes/test-bicep-recipes/extenders-aws-s3-recipe.bicep
