+++ test/functional-cloud/shared/resources/testdata/recipes/test-terraform-recipes/azure-storage/variables.tf
