+++ test/functional/datastoresrp/resources/mongodb_test.go
@@ -17,7 +17,6 @@ limitations under the License.
 package resource_test
 
 import (
-	"os"
 	"testing"
 
 	"github.com/radius-project/radius/test/functional"
@@ -69,163 +68,6 @@ func Test_MongoDB_Manual(t *testing.T) {
 						validation.NewK8sPodForResource(name, "mdb-us-app-ctnr").ValidateLabels(false),
 						validation.NewK8sPodForResource(name, "mdb-us-ctnr").ValidateLabels(false),
 						validation.NewK8sServiceForResource(name, "mdb-us-rte").ValidateLabels(false),
-					},
-				},
-			},
-		},
-	})
-
-	test.Test(t)
-}
-
-// Test_MongoDB_Recipe validates:
-// the creation of a mongoDB from a recipe that uses an Azure resource
-func Test_MongoDB_Recipe(t *testing.T) {
-	template := "testdata/datastoresrp-resources-mongodb-recipe.bicep"
-	name := "dsrp-resources-mongodb-recipe"
-	appNamespace := "dsrp-resources-mongodb-recipe-app"
-
-	test := shared.NewRPTest(t, name, []shared.TestStep{
-		{
-			Executor: step.NewDeployExecutor(template, functional.GetMagpieImage(), functional.GetBicepRecipeRegistry(), functional.GetBicepRecipeVersion()),
-			RPResources: &validation.RPResourceSet{
-				Resources: []validation.RPResource{
-					{
-						Name: "dsrp-resources-mongodb-recipe-env",
-						Type: validation.EnvironmentsResource,
-					},
-					{
-						Name: "dsrp-resources-mongodb-recipe",
-						Type: validation.ApplicationsResource,
-						App:  name,
-					},
-					{
-						Name: "mongodb-app-ctnr",
-						Type: validation.ContainersResource,
-						App:  name,
-					},
-					{
-						Name: "mongodb-db",
-						Type: validation.MongoDatabasesResource,
-						App:  name,
-					},
-				},
-			},
-			K8sObjects: &validation.K8sObjectSet{
-				Namespaces: map[string][]validation.K8sObject{
-					appNamespace: {
-						validation.NewK8sPodForResource(name, "mongodb-app-ctnr").ValidateLabels(false),
-					},
-				},
-			},
-		},
-	})
-
-	test.Test(t)
-}
-
-// Test_MongoDB_Recipe_Parameters validates the creation of a mongoDB from recipe with parameters passed by operator while linking recipe
-// and developer while creating the mongoDatabase resource.
-// If the same parameters are set by the developer and the operator then the developer parameters are applied in to resolve conflicts.
-// Container uses the mongoDB resource to connect to the mongoDB resource
-func Test_MongoDB_RecipeParameters(t *testing.T) {
-	t.Skip("Skipping test as creating/deleting cosmosdb resource is unreliable - https://github.com/radius-project/radius/issues/5929")
-
-	template := "testdata/datastoresrp-resources-mongodb-recipe-parameters.bicep"
-	name := "dsrp-resources-mongodb-recipe-parameters"
-	appNamespace := "dsrp-resources-mongodb-recipe-param-app"
-	rg := os.Getenv("INTEGRATION_TEST_RESOURCE_GROUP_NAME")
-	// Error the test if INTEGRATION_TEST_RESOURCE_GROUP_NAME is not set
-	// for running locally set the INTEGRATION_TEST_RESOURCE_GROUP_NAME with the test resourceGroup
-	if rg == "" {
-		t.Error("This test needs the env variable INTEGRATION_TEST_RESOURCE_GROUP_NAME to be set")
-	}
-
-	test := shared.NewRPTest(t, name, []shared.TestStep{
-		{
-			Executor: step.NewDeployExecutor(template, functional.GetMagpieImage(), functional.GetBicepRecipeRegistry(), functional.GetBicepRecipeVersion()),
-			RPResources: &validation.RPResourceSet{
-				Resources: []validation.RPResource{
-					{
-						Name: "dsrp-resources-env-recipe-parameters-env",
-						Type: validation.EnvironmentsResource,
-					},
-					{
-						Name: name,
-						Type: validation.ApplicationsResource,
-						App:  name,
-					},
-					{
-						Name: "mdb-param-ctnr",
-						Type: validation.ContainersResource,
-						App:  name,
-					},
-					{
-						Name: "mdb-recipe-param-db",
-						Type: validation.MongoDatabasesResource,
-						App:  name,
-					},
-				},
-			},
-			K8sObjects: &validation.K8sObjectSet{
-				Namespaces: map[string][]validation.K8sObject{
-					appNamespace: {
-						validation.NewK8sPodForResource(name, "mdb-param-ctnr").ValidateLabels(false),
-					},
-				},
-			},
-		},
-	})
-
-	test.Test(t)
-}
-
-// Test_MongoDB_Recipe_ContextParameter validates creation of a mongoDB from
-// a default recipe using the context parameter generated and set by DatastoresRP,
-// and container using the mongoDatabases portable resource to connect to the underlying mongoDB resource.
-func Test_MongoDB_Recipe_ContextParameter(t *testing.T) {
-	t.Skip("Skipping test as creating/deleting cosmosdb resource is unreliable - https://github.com/radius-project/radius/issues/5929")
-
-	template := "testdata/datastoresrp-resources-mongodb-recipe-context.bicep"
-	name := "dsrp-resources-mongodb-recipe-context"
-	appNamespace := "dsrp-resources-mongodb-recipe-context-app"
-	rg := os.Getenv("INTEGRATION_TEST_RESOURCE_GROUP_NAME")
-	// Error the test if INTEGRATION_TEST_RESOURCE_GROUP_NAME is not set
-	// for running locally set the INTEGRATION_TEST_RESOURCE_GROUP_NAME with the test resourceGroup
-	if rg == "" {
-		t.Error("This test needs the env variable INTEGRATION_TEST_RESOURCE_GROUP_NAME to be set")
-	}
-
-	test := shared.NewRPTest(t, name, []shared.TestStep{
-		{
-			Executor: step.NewDeployExecutor(template, functional.GetMagpieImage(), functional.GetBicepRecipeRegistry(), functional.GetBicepRecipeVersion()),
-			RPResources: &validation.RPResourceSet{
-				Resources: []validation.RPResource{
-					{
-						Name: "dsrp-resources-env-recipes-context-env",
-						Type: validation.EnvironmentsResource,
-					},
-					{
-						Name: name,
-						Type: validation.ApplicationsResource,
-						App:  name,
-					},
-					{
-						Name: "mdb-ctx-ctnr",
-						Type: validation.ContainersResource,
-						App:  name,
-					},
-					{
-						Name: "mdb-ctx",
-						Type: validation.MongoDatabasesResource,
-						App:  name,
-					},
-				},
-			},
-			K8sObjects: &validation.K8sObjectSet{
-				Namespaces: map[string][]validation.K8sObject{
-					appNamespace: {
-						validation.NewK8sPodForResource(name, "mdb-ctx-ctnr").ValidateLabels(false),
 					},
 				},
 			},
